[skip-dot-release] [release 1.5] fix CI issue in release branch
Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

This PR cherry picks two fixes into a single commit (as both are needed for the CI to pass). First one is ko moving to a new repo and thus we need to use a newer version of the github action. Second one is kapp has a new module which needs to be updated.