fix: Revisions stuck in restart loop now can scale down once unreachable
1. Unreachable pa becomes inactive

PA computeActiveCondition() will MarkInactive
in the case of "Queued" when the pa is
Unreachable.

2. Adding DeadStart e2e tests

TestDeadStartToHealthy: creates a service which is never able to
transition to a ready state by existing immediately.
The failed revision will scale to zero once the service
is updated with a healthy revision.

TestDeadStartFromHealthy: updates a healthy service with an image
that can never reach a ready state.
The healthy revision remains Ready and the DeadStart revision doesn't
not scale down until ProgressDeadline is reached.