feat: Add ability to configure multiple wilcard domains (#14543)
* feat: Add ability to configure multiple wilcard domains

- Update nscert reconciler to be able to create multiple wildcard
  certificates for any domain that is a wildcard
- Update domain config to allow marking domains as having wildcard
  enable.  The format of the change an implementation allows for prior
  usage without wildcard field to continue to function as it was.
- Update references to config object as necessary to use new object
  format

* fix: Correct issue with generated deepcopy file.

* refactor: Change wildcard domain to use type field

Change the boolean wildcard config to use a type field that could
support other usage later.  Update references and logic throughout.

* fix: Correct test fixture function for domainConfig

Since this function bypasses the configuration loading it needs to set
the default domain to Type of wildcard.

* refactor(nscert): Change cert cleanup behavior

- Cleanup certs which no longer match a wildcard domains.
- Return on first error instead of continuing (similar to prior
  behavior)

* refactor(nscert): Use label for findNamespaceCert

Change findNamespaceCert to use the labeled domain to find matching cert
instead of the dnsNames in the cert.  This seems to be the intended
behavior to keep the cert associated with the entry in the domain config
updated instead of potentially creating another.