Disable logstream so the Github Action Log Viewer works (#15146)
gotestsum now prints the entire test output into the github
action group. It does this for successful tests.

Because the logstream is very verbose this breaks the Github
Action UI.

The change in this PR disables the logstream because it is
already available to download through cluster logs which
are attached to the Github Action output as an artifact