Use the private service's hostname & target port
When we set 'ClusterIP: None' on the private service the activator
can't perform the fallback service probing. This is needed
while in mesh mode.

This change tweaks the fallback probing to use the hostname
of the private service instead of using a cluster IP.

On caveat is that when using a headless service Istio/K8s doesn't
perform any port translation 80 (http)->8012 (queue-proxy http).

Thus we perform this lookup ourselves.