Configure deployments to have no own revisionHistoryLimit (#15698)
The revision history limit of a deployment are its replicasets. As Knative has its own concept of revisions and the way a deployment looks like for revision is defined by the revision, there is no reason to maintain a history of replicasets. Multiple replicasets can exist if Knative configuration is updated (config-deployment) or if Knative itself is updated (and the queue-proxy image changes). Having no revisionHistoryLimit on the Deployment saves a lot of space in the Kubernetes database.

Signed-off-by: Sascha Schwarze <schwarzs@de.ibm.com>
Co-authored-by: Subhasree Das <subhasree.das@ibm.com>