Ensure configured Transport is used
If an integrator customizes the Transport in an Options function, we want to
apply the provided Transport to our default httputil.ReverseProxy
regardless of which Option sets the Transport and regardless of whether
or not other Option functions replace our default http.Handler.

Add some tests for this since the interactions beteween
Options.Transport and Options.ProxyHandler can be subtle.