Make queue-proxy's reverse proxy handler extendable (#16097)
* Make queue-proxy's reverse proxy handler extendable

This builds on #13133 to make it possible to adjust settings on the
*httputil.ReverseProxy that queue-proxy uses, or to replace it entirely
with any http.Handler, using the out-of-tree extension pattern.

* Ensure configured Transport is used

If an integrator customizes the Transport in an Options function, we want to
apply the provided Transport to our default httputil.ReverseProxy
regardless of which Option sets the Transport and regardless of whether
or not other Option functions replace our default http.Handler.

Add some tests for this since the interactions beteween
Options.Transport and Options.ProxyHandler can be subtle.