Loosen cgroupv2 cpu.weight check in test (#16147)
Our spec doesn't specify **how** the weight value has to be set. It just specifies that it MUST use them and that it SHOULD expose them.

The formula used in our test taken from a Kubernetes KEP has been subject to discussion in https://github.com/kubernetes/kubernetes/issues/131216 and runc's 1.3.2 release has implemented a new formula. Thus this test currently breaks on systems that use the newest runc release.

This changes the check to not rely on a specific formula but instead assert that the weight is a "reasonable" value bigger than 0.