Allow manual scaling of the Pod Autoscaler (#16186)
* Exclude autoscaling annotations from everything but the KPA

When creating child resources for a Revisino we now exclude
propagating autoscaling annotations. The exception is the KPA
for which we do propagate annotations.

* propagate annotation updates to the Pod Autoscaler

* Propagate the KPA status even when updating the KPA fails

Otherwise if the update fails this can have a side-effect of
marking a revision as ready.

Specifically when a Deployment has >1 ready Pods it'll mark the
Revision as Ready. It's the KPA status propagation that will mark
the revision still deploying when it hasn't reached min-scale.

* don't remove annotations that are added by defaulting/webhooks

* tweak diff output so we don't see defaulted values always being added

* handle the case when an annotation is deleted

* fix unit tests

* include e2e test

* minor refactoring - tweak func names

* further refactoring and godoc

* add a unit test to cover deleting an autoscaling annotation

* update key name to match test name

* address PR feedback